health_check_using_rpc: wrap HealthCheckTask
